fix: move all typescript helpers to a separate file - `ts-utils`
fixes #216
